Moved some files to folder application
