+++ Source/application/image_builder.rb
@@ -1,4 +1,4 @@
-require './app_section'
+require './application/app_section'
 
 module Elang
   class ImageBuilder
