+++ Source/application/relocation_initializer.rb
