+++ Source/application/relocation_item.rb
