+++ Source/application/symbol_reference.rb
