+++ Source/application/variable.rb
