--- Source/main_code.rb
@@ -1,4 +1,0 @@
-module Elang
-  class MainCode < ICodeContainer
-  end
-end
